requirements.txt
selenium==3.141.0
playsound==1.2.2
gTTS==2.2.2
kafka-python:2.0.2